pyproject.toml
[project]
name = "todo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "greenlet>=3.2.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "alembic>=1.16.4",
    "asgi-lifespan>=2.1.0",
    "httpx>=0.27",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.23",
]
